Don't sync codium plugins from WSL

+++ file-install.sh
@@ -87,24 +87,26 @@ if has_arg "codium" && [[ $(which codium) ]]; then
             fi
         fi
     done
-    EXTENSIONS=$($RUN_VSC --list-extensions | grep -v simple-vim | sed 's/\r//g')
-    if [[ "$EXTENSIONS" != "$(cat $FILES_DIR/VSCodium/extensions.txt)" ]]
-    then
-        echo "$EXTENSIONS" > $FILES_DIR/VSCodium/extensions.txt
-        echo "Local VSCodium extensions different than tracked. List updated here."
-        echo "Determine desired list of extensions and run"
-        echo "    cat system_files/VSCodium/extensions.txt | "
-        echo "        xargs -n 1 -I {} bash -c \"$RUN_VSC --install-extension \\\$1\" _ {}"
+    if ! $WSL; then
+        EXTENSIONS=$($RUN_VSC --list-extensions | grep -v simple-vim | sed 's/\r//g')
+        if [[ "$EXTENSIONS" != "$(cat $FILES_DIR/VSCodium/extensions.txt)" ]]
+        then
+            echo "$EXTENSIONS" > $FILES_DIR/VSCodium/extensions.txt
+            echo "Local VSCodium extensions different than tracked. List updated here."
+            echo "Determine desired list of extensions and run"
+            echo "    cat system_files/VSCodium/extensions.txt | "
+            echo "        xargs -n 1 -I {} bash -c \"$RUN_VSC --install-extension \\\$1\" _ {}"
+        fi
+        # simple-vim
+        URL="https://github.com$(curl -s https://github.com/kkredit/vscode-simple-vim/releases | \
+                grep simple-vim-*.*.*.vsix | grep href | cut -d\" -f2)"
+        wget -q $URL
+        mv simple-vim-?.?.?.vsix $VSC_CONF_DIR/
+        if $WSL; then
+            $RUN_VSC --install-extension $(wslpath -w $VSC_CONF_DIR/simple-vim-?.?.?.vsix) >/dev/null
+        else
+            $RUN_VSC --install-extension $VSC_CONF_DIR/simple-vim-?.?.?.vsix >/dev/null
+        fi
+        rm $VSC_CONF_DIR/simple-vim-?.?.?.vsix
     fi
-    # simple-vim
-    URL="https://github.com$(curl -s https://github.com/kkredit/vscode-simple-vim/releases | \
-            grep simple-vim-*.*.*.vsix | grep href | cut -d\" -f2)"
-    wget -q $URL
-    mv simple-vim-?.?.?.vsix $VSC_CONF_DIR/
-    if $WSL; then
-        $RUN_VSC --install-extension $(wslpath -w $VSC_CONF_DIR/simple-vim-?.?.?.vsix) >/dev/null
-    else
-        $RUN_VSC --install-extension $VSC_CONF_DIR/simple-vim-?.?.?.vsix >/dev/null
-    fi
-    rm $VSC_CONF_DIR/simple-vim-?.?.?.vsix
 fi
